Fix issue with git grep when no file extensions are specified
Currently if no file extensions are specified then git grep will not be given a
path and will therefore only search from the current directory instead of the
root. So in that case we add the magic path (:/) to search from the project
root. Note that this only works in Git versions greater then 1.7.6.